Add files via upload
This is the original CROWN model. When I reproduce this model, I wrote a lot of detailed comments for a better understanding. The simple modification that I only change word2vec to BERT is also included. However, these are not the focus of this project, so I zip them together when uploading. Those unzip files are the main part of this project and the improvements are good, we can even consider the improvement is a totally new model!